Feature/fine tune workout display

+++ src/main/java/com/sisa/tabata/ui/listener/loader/WorkoutTextViewLongClickListener.java
@@ -2,10 +2,12 @@ package com.sisa.tabata.ui.listener.loader;
 
 import android.view.View;
 import android.widget.LinearLayout;
+import android.widget.Toast;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.sisa.tabata.TabataApplication;
+import com.sisa.tabata.dao.loader.LoadedWorkoutProvider;
 import com.sisa.tabata.dao.service.DatabaseDataService;
 
 import roboguice.RoboGuice;
@@ -16,8 +18,11 @@ import roboguice.RoboGuice;
 @Singleton
 public class WorkoutTextViewLongClickListener implements View.OnLongClickListener {
 
+    public static final int MINIMUM_REMAINING_WORKOUTS_COUNT = 1;
     @Inject
     private DatabaseDataService databaseDataService;
+    @Inject
+    private LoadedWorkoutProvider loadedWorkoutProvider;
     private LinearLayout existingWorkoutLayout;
 
     public WorkoutTextViewLongClickListener() {
@@ -26,17 +31,37 @@ public class WorkoutTextViewLongClickListener implements View.OnLongClickListene
 
     @Override
     public boolean onLongClick(View view) {
-        deleteWorkout((int) view.getTag());
-        refreshWorkoutTextView(view);
+        deleteWorkout(view);
         return true;
     }
 
-    private void deleteWorkout(int id) {
-        databaseDataService.deleteWorkoutById(id);
+    private void deleteWorkout(View view) {
+        if (hasWorkoutLeftAfterDelete()) {
+            int id = (int) view.getTag();
+            databaseDataService.deleteWorkoutById(id);
+            refreshLoadedWorkout(id);
+            refreshWorkoutTextView(view);
+        } else {
+            showCantDeleteMessage();
+        }
+    }
+
+    private void showCantDeleteMessage() {
+        Toast.makeText(TabataApplication.getAppContext(), "Can't delete last workout", Toast.LENGTH_SHORT).show();
     }
 
     private void refreshWorkoutTextView(View view) {
         existingWorkoutLayout.removeView(view);
+    }
+
+    private void refreshLoadedWorkout(int id) {
+        if (loadedWorkoutProvider.getLoadedWorkout().getId() == id){
+            loadedWorkoutProvider.loadFirstWorkoutInList();
+        }
+    }
+
+    private boolean hasWorkoutLeftAfterDelete() {
+        return databaseDataService.getAllWorkoutsSortedList().size() > MINIMUM_REMAINING_WORKOUTS_COUNT;
     }
 
     public void setExistingWorkoutLayout(LinearLayout existingWorkoutLayout) {
